fix!: Hot fix for timezone daylight saving mau.se
This fix tries to know if the input date is a daylight saving time or
winter time. This fixes how the event is later laid out. The
implementation is prone to bug since I don't know if time change on the
date is fixed for every year or it changes.

NOTE: Timezone and daylight saving is a pain :)